requirements.txt
google-adk
litellm
fastapi
uvicorn[standard]
pymongo
python-dotenv
pypdf
imagekitio==4.2.0
pydantic
pydantic-settings
google-auth
google-auth-oauthlib
google-api-python-client
authlib
itsdangerous
python-jose[cryptography]
httpx
